fix(windows): corrige erros encontrados na revisão contra os headers do RNW
Sem compilador, mas os headers estão em node_modules. Conferido cada API
usada — as quatro suspeitas do README estavam erradas: JSValueType::Int64,
JSValue::Type(), ReactPromise<void> e JSValue(nullptr) todos existem.

A revisão pegou três erros prováveis:
- faltava <cstring> para _strnicmp
- faltava <stdexcept> para std::runtime_error
- sqlite3_changes() devolve int, o que tornaria a escolha de sobrecarga
  do JSValue ambígua entre bool, int64_t e double

Restam três riscos não verificáveis sem Windows: disponibilidade do
winsqlite3 no SDK, autolinking seguir a dependência file:, e o registro
em runtime no TurboModuleRegistry.

+++ modules/rnw-sqlite/windows/RnwSqlite/RnwSqliteModule.h
@@ -7,9 +7,11 @@
 #include <winsqlite/winsqlite3.h>
 
 #include <cstdlib>
+#include <cstring>    // _strnicmp
 #include <filesystem>
 #include <map>
 #include <mutex>
+#include <stdexcept>  // std::runtime_error
 #include <string>
 
 namespace winrt::RnwSqlite {
@@ -112,7 +114,9 @@ struct RnwSqliteModule {
     sqlite3_finalize(statement);
 
     React::JSValueObject result;
-    result["rowsAffected"] = sqlite3_changes(db);
+    // Cast explícito: JSValue tem construtores para bool, int64_t e double,
+    // e um `int` cru deixaria a escolha da sobrecarga ambígua.
+    result["rowsAffected"] = static_cast<double>(sqlite3_changes(db));
 
     // insertId só faz sentido em INSERT. Devolver o rowid de um DELETE
     // levaria o chamador a gravar um id que não é dele.
